Python3_HomeWork_Chat/Lesson_3_Networking Basics/Lesson_echo_chat/requirements.txt
ipython
jupyter
chardet
pyyaml
